orchestrate/requirements.txt
aiodns==1.1.1
aiohttp==3.3.2
apscheduler==3.5.3
cchardet==2.1.1
fire==0.1.3
pytest==3.5.1
pytest-asyncio==0.9.0
pytz==2017.2
pyyaml==3.13
requests==2.18.4
